Add RetrySocketConnector
This moves the retry logic into a separate RetrySocketConnector, which will retry a given number of times to connect with a delegate SocketConnector object. The connector waits some number of seconds between connection attempts, determined by the number of prior attempts and a new ConnectContext option, exponential backoff factor.

This changes connection logic to try each host returned by a DNS query on each connection attempt. The current behavior is a bit unintuitive IMO, as the the number of attempts could limit the number of hosts attempted. e.g., a query returning 4 hosts will only attempt to connect to the first 2 as DnsSocketConnector is currently.